+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.7.3 - Unreleased
 
+
+### Added
+* New `transform.py` module
+
 ### Fixed
 * Fixed numpy warning by using `float` instead of `np.float`.